requirements.txt
streamlit
sqlalchemy
pymysql
langchain
openai
langgraph
python-dotenv
Pillow
langchain-google-genai
langchain-community
fastapi
uvicorn
pydantic
typing_extensions
langchain_openai
langchain_core
uvicorn
requests
pytesseract
easyocr
google-genai
python-dotenv
SQLAlchemy
pymysql
chromadb
transformers
python-dotenv
easyocr
torch
faiss-cpu
email-validator
cloud-sql-python-connector